detect editors on windows
