+++ src/modules/user-settings/lib/editor.ts
@@ -1,5 +1,3 @@
-import * as fs from 'fs';
-import * as path from 'path';
 import { __ } from '@wordpress/i18n';
 
 export type SupportedEditor = 'vscode' | 'phpstorm' | 'cursor' | 'windsurf' | 'webstorm';
@@ -70,63 +68,3 @@ export const supportedEditorConfig: Record< SupportedEditor, SupportedEditorConf
 		],
 	},
 };
-
-/**
- * Expands Windows environment variables in a path
- */
-export function winExpandPath( path: string ): string {
-	return path.replace( /%([^%]+)%/g, ( _, n ) => process.env[ n ] || '' );
-}
-
-/**
- * Finds the executable path for a given editor on Windows
- */
-export async function winFindEditorPath( editorName: string ): Promise< string | null > {
-	const editor = supportedEditorConfig[ editorName as SupportedEditor ];
-	if ( ! editor || ! editor.winPaths ) {
-		return null;
-	}
-
-	for ( const possiblePath of editor.winPaths ) {
-		const expandedPath = winExpandPath( possiblePath );
-
-		// Handle wildcards in paths (for JetBrains Toolbox installations)
-		if ( expandedPath.includes( '*' ) ) {
-			const basePath = path.dirname( expandedPath );
-			const pattern = path.basename( expandedPath );
-
-			try {
-				const files = await fs.promises.readdir( basePath );
-				const matchingFiles = files.filter( ( f ) =>
-					f.match( new RegExp( pattern.replace( '*', '.*' ) ) )
-				);
-
-				for ( const file of matchingFiles ) {
-					const fullPath = path.join( basePath, file );
-					if (
-						await fs.promises
-							.access( fullPath )
-							.then( () => true )
-							.catch( () => false )
-					) {
-						return fullPath;
-					}
-				}
-			} catch ( error ) {
-				// Skip if directory doesn't exist
-				continue;
-			}
-		} else {
-			if (
-				await fs.promises
-					.access( expandedPath )
-					.then( () => true )
-					.catch( () => false )
-			) {
-				return expandedPath;
-			}
-		}
-	}
-
-	return null;
-}
